Include GNUInstallDirs before using variables that rely on it.
This was subtle because when building by hand, and running cmake
multiple times, it seems to cache the variable.  However when building
for Gentoo in an ebuild, cmake makes one pass, and the
CMAKE_INSTALL_DATAROOTDIR variable is empty.  This results in the
translation files going in the wrong place.