Allow the ToolPicker to have a scrollbar
The ToolPicker is never fully visible when kImageAnnotator is used embedded within an program,
like Spectacle. Which makes it almost impossible for the user to know if there are more
tools to choose from. Instead of going by pure luck/intuition and resizing the main window,
in this example the window of Spectacle, showing a scrollbar by default gives a hint to the
user that there are more tools available.